+++ RandomSkunk.Results/Operations/ToFailIf.cs
@@ -1,6 +1,6 @@
 namespace RandomSkunk.Results;
 
-/// <content> Defines the <c>FailIf</c> method. </content>
+/// <content> Defines the <c>ToFailIf</c> method. </content>
 public partial struct Result
 {
     /// <summary>
@@ -11,7 +11,7 @@ public partial struct Result
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public Result FailIf(Func<bool> predicate, Func<Error>? getError = null)
+    public Result ToFailIf(Func<bool> predicate, Func<Error>? getError = null)
     {
         if (predicate is null) throw new ArgumentNullException(nameof(predicate));
 
@@ -22,7 +22,7 @@ public partial struct Result
     }
 }
 
-/// <content> Defines the <c>FailIf</c> method. </content>
+/// <content> Defines the <c>ToFailIf</c> method. </content>
 public partial struct Result<T>
 {
     /// <summary>
@@ -33,7 +33,7 @@ public partial struct Result<T>
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public Result<T> FailIf(Func<T, bool> predicate, Func<T, Error>? getError = null)
+    public Result<T> ToFailIf(Func<T, bool> predicate, Func<T, Error>? getError = null)
     {
         if (predicate is null) throw new ArgumentNullException(nameof(predicate));
 
@@ -44,7 +44,7 @@ public partial struct Result<T>
     }
 }
 
-/// <content> Defines the <c>FailIf</c> method. </content>
+/// <content> Defines the <c>ToFailIf</c> method. </content>
 public partial struct Maybe<T>
 {
     /// <summary>
@@ -55,7 +55,7 @@ public partial struct Maybe<T>
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public Maybe<T> FailIf(Func<T, bool> predicate, Func<T, Error>? getError = null)
+    public Maybe<T> ToFailIf(Func<T, bool> predicate, Func<T, Error>? getError = null)
     {
         if (predicate is null) throw new ArgumentNullException(nameof(predicate));
 
@@ -66,7 +66,7 @@ public partial struct Maybe<T>
     }
 }
 
-/// <content> Defines the <c>FailIf</c> extension methods. </content>
+/// <content> Defines the <c>ToFailIf</c> extension methods. </content>
 public static partial class ResultExtensions
 {
     /// <summary>
@@ -78,8 +78,8 @@ public static partial class ResultExtensions
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public static async Task<Result> FailIf(this Task<Result> sourceResult, Func<bool> predicate, Func<Error>? getError = null) =>
-        (await sourceResult.ConfigureAwait(false)).FailIf(predicate, getError);
+    public static async Task<Result> ToFailIf(this Task<Result> sourceResult, Func<bool> predicate, Func<Error>? getError = null) =>
+        (await sourceResult.ConfigureAwait(false)).ToFailIf(predicate, getError);
 
     /// <summary>
     /// Gets a <c>Fail</c> result if <paramref name="predicate"/> returns <see langword="true"/> and this is a <c>Success</c>
@@ -91,8 +91,8 @@ public static partial class ResultExtensions
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public static async Task<Result<T>> FailIf<T>(this Task<Result<T>> sourceResult, Func<T, bool> predicate, Func<T, Error>? getError = null) =>
-        (await sourceResult.ConfigureAwait(false)).FailIf(predicate, getError);
+    public static async Task<Result<T>> ToFailIf<T>(this Task<Result<T>> sourceResult, Func<T, bool> predicate, Func<T, Error>? getError = null) =>
+        (await sourceResult.ConfigureAwait(false)).ToFailIf(predicate, getError);
 
     /// <summary>
     /// Gets a <c>Fail</c> result if <paramref name="predicate"/> returns <see langword="true"/> and this is a <c>Success</c>
@@ -104,6 +104,6 @@ public static partial class ResultExtensions
     /// <param name="getError">An optional function that gets the <c>Fail</c> result's error.</param>
     /// <returns>A <c>Fail</c> result if <paramref name="predicate"/> returned <see langword="true"/>, or the same result if it
     ///     did not.</returns>
-    public static async Task<Maybe<T>> FailIf<T>(this Task<Maybe<T>> sourceResult, Func<T, bool> predicate, Func<T, Error>? getError = null) =>
-        (await sourceResult.ConfigureAwait(false)).FailIf(predicate, getError);
+    public static async Task<Maybe<T>> ToFailIf<T>(this Task<Maybe<T>> sourceResult, Func<T, bool> predicate, Func<T, Error>? getError = null) =>
+        (await sourceResult.ConfigureAwait(false)).ToFailIf(predicate, getError);
 }